+++ astropy/units/core.py
@@ -9,10 +9,8 @@ from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 from ..extern import six
 
-import contextlib
 import inspect
 import numbers
-import re
 import sys
 import textwrap
 import warnings
@@ -32,9 +30,10 @@ __all__ = [
     'UnitsError', 'UnitsException', 'UnitsWarning', 'UnitBase',
     'NamedUnit', 'IrreducibleUnit', 'Unit', 'def_unit',
     'CompositeUnit', 'PrefixUnit', 'UnrecognizedUnit',
-    'get_current_unit_registry', 'set_enabled_units',
-    'add_enabled_units', 'set_enabled_units_context',
-    'add_enabled_units_context', 'dimensionless_unscaled']
+    'get_current_unit_registry',
+    'set_enabled_units', 'add_enabled_units',
+    'set_enabled_equivalencies', 'add_enabled_equivalencies',
+    'dimensionless_unscaled']
 
 
 def _flatten_units_collection(items):
@@ -66,19 +65,70 @@ def _flatten_units_collection(items):
     return result
 
 
+def _normalize_equivalencies(equivalencies):
+    """
+    Normalizes equivalencies, ensuring each is a 4-tuple of the form::
+
+    (from_unit, to_unit, forward_func, backward_func)
+
+    Parameters
+    ----------
+    equivalencies : list of equivalency pairs
+
+    Raises
+    ------
+    ValueError if an equivalency cannot be interpreted
+    """
+    if equivalencies is None:
+        return []
+
+    normalized = []
+
+    for i, equiv in enumerate(equivalencies):
+        if len(equiv) == 2:
+            funit, tunit = equiv
+            a = b = lambda x: x
+        elif len(equiv) == 3:
+            funit, tunit, a = equiv
+            b = a
+        elif len(equiv) == 4:
+            funit, tunit, a, b = equiv
+        else:
+            raise ValueError(
+                "Invalid equivalence entry {0}: {1!r}".format(i, equiv))
+        if not (isinstance(funit, UnitBase) and
+                (isinstance(tunit, UnitBase) or tunit is None) and
+                six.callable(a) and
+                six.callable(b)):
+            raise ValueError(
+                "Invalid equivalence entry {0}: {1!r}".format(i, equiv))
+        normalized.append((funit, tunit, a, b))
+
+    return normalized
+
+
 class _UnitRegistry(object):
     """
     Manages a registry of the enabled units.
     """
-    def __init__(self, init=[]):
-        self._reset()
-        self.add_enabled_units(init)
+    def __init__(self, init=[], equivalencies=[]):
+        if isinstance(init, _UnitRegistry):
+            equivalencies = init.equivalencies
+            init = init.all_units
 
-    def _reset(self):
+        self._reset_units()
+        self._reset_equivalencies()
+        self.add_enabled_units(init)
+        self.add_enabled_equivalencies(equivalencies)
+
+    def _reset_units(self):
         self._all_units = set()
         self._non_prefix_units = set()
         self._registry = {}
         self._by_physical_type = {}
+
+    def _reset_equivalencies(self):
+        self._equivalencies = set()
 
     @property
     def registry(self):
@@ -108,7 +158,7 @@ class _UnitRegistry(object):
             methods like `UnitBase.find_equivalent_units` and
             `UnitBase.compose`.
         """
-        self._reset()
+        self._reset_units()
         return self.add_enabled_units(units)
 
     def add_enabled_units(self, units):
@@ -162,38 +212,90 @@ class _UnitRegistry(object):
             unit._get_physical_type_id(),
             set())
 
+    @property
+    def equivalencies(self):
+        return list(self._equivalencies)
 
-_current_unit_registry = _UnitRegistry()
+    def set_enabled_equivalencies(self, equivalencies):
+        """
+        Sets the equivalencies enabled in the unit registry.
+
+        These equivalencies are used if no explicit equivalencies are given,
+        both in unit conversion and in finding equivalent units.
+
+        This is meant in particular for allowing angles to be dimensionless.
+        Use with care.
+
+        Parameters
+        ----------
+        equivalencies : list of equivalent pairs
+            E.g., as returned by `astropy.units.angles_dimensionless`.
+        """
+        self._reset_equivalencies()
+        return self.add_enabled_equivalencies(equivalencies)
+
+    def add_enabled_equivalencies(self, equivalencies):
+        """
+        Adds to the set of equivalencies enabled in the unit registry.
+
+        These equivalencies are used if no explicit equivalencies are given,
+        both in unit conversion and in finding equivalent units.
+
+        This is meant in particular for allowing angles to be dimensionless.
+        Use with care.
+
+        Parameters
+        ----------
+        equivalencies : list of equivalent pairs
+            E.g., as returned by `astropy.units.angles_dimensionless`.
+        """
+        # pre-normalize list to help catch mistakes
+        equivalencies = _normalize_equivalencies(equivalencies)
+        self._equivalencies |= set(equivalencies)
+
+
+class _UnitContext(object):
+    def __init__(self, init=[], equivalencies=[]):
+        _unit_registries.append(_UnitRegistry(init=init,
+                                              equivalencies=equivalencies))
+
+    def __enter__(self):
+        pass
+
+    def __exit__(self, type, value, tb):
+        _unit_registries.pop()
+
+
+_unit_registries = [_UnitRegistry()]
+
+
 def get_current_unit_registry():
-    return _current_unit_registry
+    return _unit_registries[-1]
 
 
 def set_enabled_units(units):
-    return get_current_unit_registry().set_enabled_units(units)
-set_enabled_units.__doc__ = _UnitRegistry.set_enabled_units.__doc__
-
-
-def add_enabled_units(units):
-    return get_current_unit_registry().add_enabled_units(units)
-add_enabled_units.__doc__ = _UnitRegistry.add_enabled_units.__doc__
-
-
-@contextlib.contextmanager
-def set_enabled_units_context(units):
     """
-    A context manager for temporarily setting the set of units that
-    are searched by methods like `UnitBase.find_equivalent_units`.
+    Sets the units enabled in the unit registry.
+
+    These units are searched when using
+    `UnitBase.find_equivalent_units`, for example.
+
+    This may be used either permanently, or as a context manager using
+    the ``with`` statement (see example below).
 
     Parameters
     ----------
     units : list of sequences, dicts, or modules containing units, or units
-        See `set_enabled_units`.
+        This is a list of things in which units may be found
+        (sequences, dicts or modules), or units themselves.  The
+        entire set will be "enabled" for searching through by methods
+        like `UnitBase.find_equivalent_units` and `UnitBase.compose`.
 
     Examples
     --------
 
     >>> from astropy import units as u
-    >>> with u.set_enabled_units_context([u.pc]):
+    >>> with u.set_enabled_units([u.pc]):
     ...     u.m.find_equivalent_units()
     ...
       Primary name | Unit definition | Aliases
@@ -213,31 +315,39 @@ def set_enabled_units_context(units):
       solRad       | 6.95508e+08 m   | R_sun, Rsun  ,
     ]
     """
-    global _current_unit_registry
-    old_registry = get_current_unit_registry()
-    _current_unit_registry = _UnitRegistry()
-    set_enabled_units(units)
-    yield
-    _current_unit_registry = old_registry
+    # get a context with a new registry, using equivalencies of the current one
+    context = _UnitContext(
+        equivalencies=get_current_unit_registry().equivalencies)
+    # in this new current registry, enable the units requested
+    get_current_unit_registry().set_enabled_units(units)
+    return context
 
 
-@contextlib.contextmanager
-def add_enabled_units_context(units):
+def add_enabled_units(units):
     """
-    A context manager for temporarily adding to the set of units that
-    are searched by methods like `UnitBase.find_equivalent_units`.
+    Adds to the set of units enabled in the unit registry.
+
+    These units are searched when using
+    `UnitBase.find_equivalent_units`, for example.
+
+    This may be used either permanently, or as a context manager using
+    the ``with`` statement (see example below).
 
     Parameters
     ----------
     units : list of sequences, dicts, or modules containing units, or units
-        See `add_enabled_units`.
+        This is a list of things in which units may be found
+        (sequences, dicts or modules), or units themselves.  The
+        entire set will be added to the "enabled" set for searching
+        through by methods like `UnitBase.find_equivalent_units` and
+        `UnitBase.compose`.
 
     Examples
     --------
 
     >>> from astropy import units as u
     >>> from astropy.units import imperial
-    >>> with u.add_enabled_units_context(imperial):
+    >>> with u.add_enabled_units(imperial):
     ...     u.m.find_equivalent_units()
     ...
       Primary name | Unit definition | Aliases
@@ -257,13 +367,67 @@ def add_enabled_units_context(units):
       yd           | 0.9144 m        | yard             ,
     ]
     """
-    global _current_unit_registry
-    old_registry = get_current_unit_registry()
-    _current_unit_registry = _UnitRegistry()
-    add_enabled_units(old_registry.all_units)
-    add_enabled_units(units)
-    yield
-    _current_unit_registry = old_registry
+    # get a context with a new registry, which is a copy of the current one
+    context = _UnitContext(get_current_unit_registry())
+    # in this new current registry, enable the further units requested
+    get_current_unit_registry().add_enabled_units(units)
+    return context
+
+
+def set_enabled_equivalencies(equivalencies):
+    """
+    Sets the equivalencies enabled in the unit registry.
+
+    These equivalencies are used if no explicit equivalencies are given,
+    both in unit conversion and in finding equivalent units.
+
+    This is meant in particular for allowing angles to be dimensionless.
+    Use with care.
+
+    Parameters
+    ----------
+    equivalencies : list of equivalent pairs
+        E.g., as returned by `astropy.units.angles_dimensionless`.
+
+    Examples
+    --------
+    Exponentiation normally requires dimensionless quantities.  To avoid
+    problems with complex phases::
+
+        >>> from astropy import units as u
+        >>> with u.set_enabled_equivalencies(u.dimensionless_angles()):
+        ...     phase = 0.5 * u.cycle
+        ...     np.exp(1j*phase)                          # doctest: +ELLIPSIS
+        <Quantity (-1+...j) >
+    """
+    # get a context with a new registry, using all units of the current one
+    context = _UnitContext(get_current_unit_registry().all_units)
+    # in this new current registry, enable the equivalencies requested
+    get_current_unit_registry().set_enabled_equivalencies(equivalencies)
+    return context
+
+
+def add_enabled_equivalencies(equivalencies):
+    """
+    Adds to the equivalencies enabled in the unit registry.
+
+    These equivalencies are used if no explicit equivalencies are given,
+    both in unit conversion and in finding equivalent units.
+
+    This is meant in particular for allowing angles to be dimensionless.
+    Since no equivalencies are enabled by default, generally it is recommended
+    to use `set_enabled_equivalencies`.
+
+    Parameters
+    ----------
+    equivalencies : list of equivalent pairs
+        E.g., as returned by `astropy.units.angles_dimensionless`.
+    """
+    # get a context with a new registry, which is a copy of the current one
+    context = _UnitContext(get_current_unit_registry())
+    # in this new current registry, enable the further equivalencies requested
+    get_current_unit_registry().add_enabled_equivalencies(equivalencies)
+    return context
 
 
 class UnitsError(Exception):
@@ -386,33 +550,28 @@ class UnitBase(object):
     @staticmethod
     def _normalize_equivalencies(equivalencies):
         """
-        Normalizes a list of equivalencies, by ensuring each element
-        is a 4-tuple of the form::
+        Normalizes equivalencies, ensuring each is a 4-tuple of the form::
 
-            (from_unit, to_unit, forward_func, backward_func)
+        (from_unit, to_unit, forward_func, backward_func)
 
-        Raises a ValueError if the equivalencies list is invalid.
+        Parameters
+        ----------
+        equivalencies : list of equivalency pairs, or `None`
+
+        Returns
+        -------
+        A normalized list, including possible global defaults set by, e.g.,
+        `set_enabled_equivalencies`, except when `equivalencies`=`None`,
+        in which case the returned list is always empty.
+
+        Raises
+        ------
+        ValueError if an equivalency cannot be interpreted
         """
-        normalized = []
-        for i, equiv in enumerate(equivalencies):
-            if len(equiv) == 2:
-                funit, tunit = equiv
-                a = b = lambda x: x
-            elif len(equiv) == 3:
-                funit, tunit, a = equiv
-                b = a
-            elif len(equiv) == 4:
-                funit, tunit, a, b = equiv
-            else:
-                raise ValueError(
-                    "Invalid equivalence entry {0}: {1!r}".format(i, equiv))
-            if not (isinstance(funit, UnitBase) and
-                    isinstance(tunit, UnitBase) and
-                    six.callable(a) and
-                    six.callable(b)):
-                raise ValueError(
-                    "Invalid equivalence entry {0}: {1!r}".format(i, equiv))
-            normalized.append((funit, tunit, a, b))
+        normalized = _normalize_equivalencies(equivalencies)
+        if equivalencies is not None:
+            normalized += get_current_unit_registry().equivalencies
+
         return normalized
 
     def _validate_power(self, p):
@@ -533,12 +692,16 @@ class UnitBase(object):
         Parameters
         ----------
         other : unit object or string or tuple
-           The unit to convert to. If a tuple of units is specified, this
-           method returns true if the unit matches any of the ones in the tuple.
+            The unit to convert to. If a tuple of units is specified, this
+            method returns true if the unit matches any of those in the tuple;
+            for this case, equivalencies are ignored.
 
         equivalencies : list of equivalence pairs, optional
-           A list of equivalence pairs to try if the units are not
-           directly convertible.  See :ref:`unit_equivalencies`.
+            A list of equivalence pairs to try if the units are not
+            directly convertible.  See :ref:`unit_equivalencies`.
+            This list is in addition to possible global defaults set by, e.g.,
+            `set_enabled_equivalencies`.
+            Use `None` to turn off all equivalencies.
 
         Returns
         -------
@@ -546,12 +709,21 @@ class UnitBase(object):
         """
 
         if isinstance(other, tuple):
-            return any(self.is_equivalent(u) for u in other)
+            return any(self.is_equivalent(u, equivalencies=None)
+                       for u in other)
         else:
             other = Unit(other, parse_strict='silent')
 
         equivalencies = self._normalize_equivalencies(equivalencies)
 
+        return self._is_equivalent(other, equivalencies)
+
+    def _is_equivalent(self, other, equivalencies=[]):
+        """Returns True if this unit is equivalent to `other`.
+        See `is_equivalent`, except that a proper Unit object should be
+        given (i.e., no string) and that the equivalency list should be
+        normalized using `_normalize_equivalencies`.
+        """
         if isinstance(other, UnrecognizedUnit):
             return False
 
@@ -562,11 +734,19 @@ class UnitBase(object):
             unit = self.decompose()
             other = other.decompose()
             for a, b, forward, backward in equivalencies:
-                if (unit.is_equivalent(a) and other.is_equivalent(b)):
-                    return True
-                elif (unit.is_equivalent(b) and other.is_equivalent(a)):
-                    return True
-            return False
+                if b is None:
+                    # after canceling, is what's left convertable
+                    # to dimensionless (according to the equivalency)?
+                    try:
+                        (unit/other).decompose([a])
+                        return True
+                    except:
+                        pass
+                else:
+                    if(unit._is_equivalent(a) and other._is_equivalent(b) or
+                       unit._is_equivalent(b) and other._is_equivalent(a)):
+                        return True
+
         return False
 
     def _apply_equivalences(self, unit, other, equivalencies):
@@ -586,14 +766,22 @@ class UnitBase(object):
         other = other.decompose()
 
         for funit, tunit, a, b in equivalencies:
-            if (unit.is_equivalent(funit) and other.is_equivalent(tunit)):
-                scale1 = (unit / funit)._dimensionless_constant()
-                scale2 = (tunit / other)._dimensionless_constant()
-                return make_converter(scale1, a, scale2)
-            elif (other.is_equivalent(funit) and unit.is_equivalent(tunit)):
-                scale1 = (unit / tunit)._dimensionless_constant()
-                scale2 = (funit / other)._dimensionless_constant()
-                return make_converter(scale1, b, scale2)
+            if tunit is None:
+                try:
+                    ratio_in_funit = (unit/other).decompose([funit])
+                    return make_converter(ratio_in_funit.scale, a, 1.)
+                except:
+                    pass
+            else:
+                if unit._is_equivalent(funit) and other._is_equivalent(tunit):
+                    scale1 = (unit / funit)._dimensionless_constant()
+                    scale2 = (tunit / other)._dimensionless_constant()
+                    return make_converter(scale1, a, scale2)
+                elif (other._is_equivalent(funit) and
+                      unit._is_equivalent(tunit)):
+                    scale1 = (unit / tunit)._dimensionless_constant()
+                    scale2 = (funit / other)._dimensionless_constant()
+                    return make_converter(scale1, b, scale2)
 
         def get_err_str(unit):
             unit_str = unit.to_string('unscaled')
@@ -620,11 +808,14 @@ class UnitBase(object):
         Parameters
         ----------
         other : unit object or string
-           The unit to convert to.
+            The unit to convert to.
 
         equivalencies : list of equivalence pairs, optional
-           A list of equivalence pairs to try if the units are not
-           directly convertible.  See :ref:`unit_equivalencies`.
+            A list of equivalence pairs to try if the units are not
+            directly convertible.  See :ref:`unit_equivalencies`.
+            This list is in addition to possible global defaults set by, e.g.,
+            `set_enabled_equivalencies`.
+            Use `None` to turn off all equivalencies.
 
         Returns
         -------
@@ -658,13 +849,16 @@ class UnitBase(object):
         other : unit object or string
             The unit to convert to.
 
-        value : scalar int or float, or sequence that can be converted to array, optional
+        value : scalar int or float, or sequence convertable to array, optional
             Value(s) in the current unit to be converted to the
             specified unit.  If not provided, defaults to 1.0
 
         equivalencies : list of equivalence pairs, optional
-           A list of equivalence pairs to try if the units are not
-           directly convertible.  See :ref:`unit_equivalencies`.
+            A list of equivalence pairs to try if the units are not
+            directly convertible.  See :ref:`unit_equivalencies`.
+            This list is in addition to possible global defaults set by, e.g.,
+            `set_enabled_equivalencies`.
+            Use `None` to turn off all equivalencies.
 
         Returns
         -------
@@ -742,10 +936,11 @@ class UnitBase(object):
         # Make a list including all of the equivalent units
         units = [unit]
         for funit, tunit, a, b in equivalencies:
-            if self.is_equivalent(funit):
-                units.append(tunit.decompose())
-            elif self.is_equivalent(tunit):
-                units.append(funit.decompose())
+            if tunit is not None:
+                if self._is_equivalent(funit):
+                    units.append(tunit.decompose())
+                elif self._is_equivalent(tunit):
+                    units.append(funit.decompose())
 
         # Store partial results
         partial_results = []
@@ -766,7 +961,7 @@ class UnitBase(object):
                 # without needing to do an exhaustive search.
                 if len(tunit_decomposed.bases) == 1:
                     for base, power in zip(u._bases, u._powers):
-                        if tunit_decomposed.is_equivalent(base):
+                        if tunit_decomposed._is_equivalent(base):
                             tunit = tunit ** power
                             tunit_decomposed = tunit_decomposed ** power
                             break
@@ -843,6 +1038,9 @@ class UnitBase(object):
         equivalencies : list of equivalence pairs, optional
             A list of equivalence pairs to also list.  See
             :ref:`unit_equivalencies`.
+            This list is in addition to possible global defaults set by, e.g.,
+            `set_enabled_equivalencies`.
+            Use `None` to turn off all equivalencies.
 
         units : set of units to compose to, optional
             If not provided, any known units may be used to compose
@@ -1018,12 +1216,13 @@ class UnitBase(object):
         unit_registry = get_current_unit_registry()
         units = set(unit_registry.get_units_with_physical_type(self))
         for funit, tunit, a, b in equivalencies:
-            if funit not in units:
-                units.update(
-                    unit_registry.get_units_with_physical_type(funit))
-            if tunit not in units:
-                units.update(
-                    unit_registry.get_units_with_physical_type(tunit))
+            if tunit is not None:
+                if funit not in units:
+                    units.update(
+                        unit_registry.get_units_with_physical_type(funit))
+                if tunit not in units:
+                    units.update(
+                        unit_registry.get_units_with_physical_type(tunit))
         return units
 
     class EquivalentUnitsList(list):
@@ -1071,6 +1270,8 @@ class UnitBase(object):
         equivalencies : list of equivalence pairs, optional
             A list of equivalence pairs to also list.  See
             :ref:`unit_equivalencies`.
+            Any list given, including an empty one, supercedes global defaults
+            that may be in effect (as set by `set_enabled_equivalencies`)
 
         units : set of units to search in, optional
             If not provided, all defined units will be searched for
@@ -1313,7 +1514,8 @@ class IrreducibleUnit(NamedUnit):
     def decompose(self, bases=set()):
         if len(bases) and not self in bases:
             for base in bases:
-                if self.is_equivalent(base):
+                # to avoid roundrip, ensure no default equivalencies get used
+                if self.is_equivalent(base, equivalencies=[]):
                     return CompositeUnit(self.to(base), [base], [1])
 
             raise UnitsError(
@@ -1356,8 +1558,8 @@ class UnrecognizedUnit(IrreducibleUnit):
             "with it are invalid.".format(self.name))
 
     __pow__ = __div__ = __rdiv__ = __truediv__ = __rtruediv__ = __mul__ = \
-               __rmul__ = __lt__ = __gt__ = __le__ = __ge__ = __neg__ = \
-               _unrecognized_operator
+        __rmul__ = __lt__ = __gt__ = __le__ = __ge__ = __neg__ = \
+        _unrecognized_operator
 
     def __eq__(self, other):
         other = Unit(other, parse_strict='silent')
@@ -1366,11 +1568,11 @@ class UnrecognizedUnit(IrreducibleUnit):
     def __ne__(self, other):
         return not (self == other)
 
-    def is_equivalent(self, other, equivalencies=[]):
+    def is_equivalent(self, other, equivalencies=None):
         self._normalize_equivalencies(equivalencies)
         return self == other
 
-    def get_converter(self, other, equivalencies=[]):
+    def get_converter(self, other, equivalencies=None):
         self._normalize_equivalencies(equivalencies)
         raise ValueError(
             "The unit {0!r} is unrecognized.  It can not be converted "
@@ -1440,15 +1642,15 @@ class _UnitMetaClass(type):
                 if parse_strict == 'silent':
                     pass
                 else:
-                    msg = "'{0}' did not parse as unit format '{1}': {2}".format(
-                        s, format, str(e))
+                    msg = ("'{0}' did not parse as unit format '{1}': {2}"
+                           .format(s, format, str(e)))
                     if parse_strict == 'raise':
                         raise ValueError(msg)
                     elif parse_strict == 'warn':
                         warnings.warn(msg, UnitsWarning)
                     else:
-                        raise ValueError(
-                            "'parse_strict' must be 'warn', 'raise' or 'silent'")
+                        raise ValueError("'parse_strict' must be 'warn', "
+                                         "'raise' or 'silent'")
                 return UnrecognizedUnit(s)
 
         elif isinstance(s, (int, float, np.floating, np.integer)):
@@ -1653,7 +1855,7 @@ class CompositeUnit(UnitBase):
         def add_unit(unit, power, scale):
             if unit not in bases:
                 for base in bases:
-                    if unit.is_equivalent(base):
+                    if unit._is_equivalent(base):
                         scale *= unit.to(base) ** power
                         unit = base
                         break